add final touches
