--- myapp/src/components/timelineAnimation.svelte
@@ -1,7 +1,0 @@
-<script>
-  import gsap from 'gsap';
-
-
-</script>
-
-
